refactored RentService, improved RentController, added sending notification from server, added checking if user can send notification or do rent

+++ MyToolsYourToolsBackend/MyToolsYourToolsBackend.API/Controllers/RentController.cs
@@ -18,17 +18,20 @@ namespace MyToolsYourToolsBackend.API.Controllers
     {
         private IRentService _rentService;
         private IUserService _userService;
+        private IOfferService _offerService;
 
-        public RentsController(IRentService rentService, IUserService userService)
+        public RentsController(IRentService rentService, IUserService userService, IOfferService offerService)
         {
             _rentService = rentService;
             _userService = userService;
+            _offerService = offerService;
         }
 
         [HttpPost("rents")]
         public IActionResult AddRent([FromBody]RentForCreationDto rentFromBody)
         {
-            if (!_userService.CheckIfUserExists(rentFromBody.BorrowerId))
+            if (!_userService.CheckIfUserExists(rentFromBody.BorrowerId)
+                || !_offerService.CheckIfOfferExists(rentFromBody.OfferId))
             {
                 return NotFound();
             }
@@ -40,6 +43,11 @@ namespace MyToolsYourToolsBackend.API.Controllers
                 return BadRequest("Niewystarczająca ilość punktów na zrealizowanie wypożyczenia.");
             }
 
+            if (!_offerService.CheckIfOfferIsActive(rentFromBody.OfferId))
+            {
+                return BadRequest("Oferta nieaktywna.");
+            }
+
             _rentService.AddRent(rentFromBody, pointsRentCost);
 
             return NoContent();
@@ -48,8 +56,15 @@ namespace MyToolsYourToolsBackend.API.Controllers
         [HttpDelete("rents/{offerId}")]
         public IActionResult DeleteRent(Guid offerId)
         {
+            if(!_offerService.CheckIfOfferExists(offerId))
+            {
+                return NotFound();
+            }
+
             int pointsReturnReward = 100;
+
             _rentService.DeleteRent(offerId, pointsReturnReward);
+
             return NoContent();
         }
     }
